requirements.txt
opencv-python
openai
ipython
python-dotenv
requests
Pillow
imutils
streamlit